component graph

+++ prisma/migrations/20240318023509_psf_to_int/migration.sql
@@ -3,7 +3,7 @@ PRAGMA foreign_keys=OFF;
 CREATE TABLE "new_PriceReport" (
     "id" INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
     "price" INTEGER NOT NULL,
-    "psf" REAL NOT NULL DEFAULT 0.00,
+    "psf" INTEGER NOT NULL DEFAULT 0,
     "date" DATETIME NOT NULL,
     "apartmentId" INTEGER NOT NULL,
     CONSTRAINT "PriceReport_apartmentId_fkey" FOREIGN KEY ("apartmentId") REFERENCES "Apartment" ("id") ON DELETE RESTRICT ON UPDATE CASCADE
